Fix a form of "auto DoS attack"
If a `poll_online()` would fail, that is, if a time transaction
(that we use for "polling") failed, we would call `_test_connection`
_and_ start a timer for (next) poll. Since `_test_connection` starts
another timer for another test with the time transaction, we would, if
connection to the Pubnub network is lost, keep piling up these time
transactions in a linear fashion. That is, each failed `poll_online()`
would start a new "thread" of time transactions, each of which would
only stop once the time transaction actually succeeds.

Basically, we don't need to react to a failed poll_online(), as it
might be a transient failure. In general, it would be nice to somehow
note that "currently, we don't seem to have a connection to Pubnub
network", but, the code doesn't support such a notion or notification.
We might add such a thing, but, it's out of scope of fixing this bug.